io_uring: be smarter about handling IOPOLL completions
When a request is polled successfully, the callback handler marks it as
such by setting req->iopoll_completed. The polling side in turns will
find these requests by browsing the entire pending list - first in a
polling loop, and subsequently in a second loop to complete requests.
If a lot of IO is pending, this can waste CPU cycles.

Rather than store the completion state in a variable in the io_kiocb,
add a 2nd iopoll list that simply stores completed requests. This can
be accessed locklessly on the completion side. With that, the 2nd
iopoll completion loop need only iterate over requests that it knows
have been completed.

Link: https://lore.kernel.org/io-uring/20251210085501.84261-1-changfengnan@bytedance.com/
Signed-off-by: Jens Axboe <axboe@kernel.dk>