io_uring: move around a few struct io_kiocb members
With the removal of ->iopoll_completed, there are a few gaps. Move the
->cancel_seq_set member to the new 1 byte hole, which then frees up the
extra needed byte near ->refs which ->poll_refs can now fill. This
shrunks struct io_kiocb from 248 to 240 bytes. The latter doesn't really
matter as it's cacheline aligned anyway, but it does make it easier if a
new member were to be added.

Signed-off-by: Jens Axboe <axboe@kernel.dk>